Move FunctionExpression to expression package
