+++ java-vtl-script/src/main/java/no/ssb/vtl/script/visitors/ExpressionVisitor.java
@@ -33,8 +33,8 @@ import no.ssb.vtl.parser.VTLParser;
 import no.ssb.vtl.script.VTLDataset;
 import no.ssb.vtl.script.error.ContextualRuntimeException;
 import no.ssb.vtl.script.error.VTLRuntimeException;
-import no.ssb.vtl.script.functions.FunctionExpression;
 import no.ssb.vtl.script.expressions.IfThenElseExpression;
+import no.ssb.vtl.script.expression.FunctionExpression;
 import no.ssb.vtl.script.functions.VTLAddition;
 import no.ssb.vtl.script.functions.VTLAnd;
 import no.ssb.vtl.script.functions.VTLConcatenation;
